Refactoring location for device id endpoint.
